requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
boto3==1.38.10
botocore==1.38.10
Django==5.2.1
django-ninja==1.4.1
django-storages==1.14.6
GDAL==3.10.3
jmespath==1.0.1
pillow==11.2.1
psycopg==3.2.7
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
django-cors-headers==4.3.1
gunicorn
whitenoise
django-jazzmin==3.0.1
requests~=2.32.3
django-celery-beat~=2.8.1
celery~=5.2.3